Move email package to model.email
