+++ src/main/java/seedu/club/commons/events/ui/SendEmailRequestEvent.java
@@ -1,9 +1,9 @@
 package seedu.club.commons.events.ui;
 
 import seedu.club.commons.events.BaseEvent;
-import seedu.club.logic.commands.email.Body;
-import seedu.club.logic.commands.email.Client;
-import seedu.club.logic.commands.email.Subject;
+import seedu.club.model.email.Body;
+import seedu.club.model.email.Client;
+import seedu.club.model.email.Subject;
 
 //@@author yash-chowdhary
 /**
